requirements.txt
bidict==0.21.4
boto3==1.20.44
botocore==1.23.44
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
Flask==2.0.2
flask-sock==0.4.0
Flask-SocketIO==5.1.1
h11==0.12.0
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
pycparser==2.21
pycryptodome==3.12.0
python-dateutil==2.8.2
python-engineio==4.3.0
python-socketio==5.5.0
s3transfer==0.5.0
simple-websocket==0.5.0
six==1.16.0
sock==0.6.0
urllib3==1.26.8
websocket-client==0.57.0
Werkzeug==2.0.2
wsproto==1.0.0